[WIP] Major refactor: Rename user_flow to policy
